Update AABB to be Vector2<f32> min / max

+++ src/components/size.rs
@@ -2,8 +2,8 @@ use amethyst::ecs::{Component, VecStorage};
 
 #[derive(Default)]
 pub struct Size {
-    pub x: f32,
-    pub y: f32,
+    pub x: u32,
+    pub y: u32,
 }
 impl Component for Size {
     type Storage = VecStorage<Self>;
